# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WordForge API
  description: |-
    WordForge 后端 API — v1-stable 端点。

    路径基础路径 `/api`（见 servers[0].url）。
    本规格由 `src/openapi.rs` 自动生成，通过 `cargo test --test openapi_export` 导出，CI 以 `git diff --exit-code docs/openapi.yaml` 防止规格漂移（drift）。
  version: 1.3.0-beta.1
servers:
- url: /api
paths:
...
  /records:
    get:
      tags:
      - learning
      summary: 获取学习记录列表
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: perPage
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: 分页学习记录
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
    post:
      tags:
      - learning
      summary: 提交学习记录
      description: 记录一次单词作答。`recordType` 缺省落库 `all`；`selfRating` 可选。`clientRecordId` 强烈建议携带并在重试时复用同一个值——服务端以此为幂等键，缺省时重试/离线重放会被当成两次独立事件二次计入 AMAS 状态。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
        required: true
      responses:
        '200':
          description: '命中 clientRecordId 幂等去重，返回原记录（`duplicate: true`）'
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '201':
          description: 已创建学习记录
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '202':
          description: 'v1.3.0 起默认（RECORDS_OUTBOX_ASYNC=true）的异步落库确认。裸 JSON 对象（非 {success,data} 信封）：`{accepted: true, async: true, clientRecordId: string}`（缺省时服务端生成），不含 record/amasResult 字段，响应形状与 200/201 不同。设 RECORDS_OUTBOX_ASYNC=false 回退同步老路。'
          content:
            application/json:
              schema:
                type: object
                required:
                - accepted
                - async
                - clientRecordId
                properties:
                  accepted:
                    type: boolean
                  async:
                    type: boolean
                  clientRecordId:
                    type: string
        '400':
          description: 请求体校验失败 — VALIDATION_ERROR / INVALID_REQUEST_BODY
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    CreateRecordRequest:
      type: object
      required:
      - wordId
      - isCorrect
      - responseTimeMs
      properties:
        clientRecordId:
          type: string
        isCorrect:
          type: boolean
        recordType:
          type: string
          enum:
          - learning
          - review
          - all
        responseTimeMs:
          type: integer
        selfRating:
          type: integer
        sessionId:
          type: string
          format: uuid
        wordId:
          type: string
          format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...